requirements.txt
# TruthLens Core API Dependencies
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
pydantic>=2.6.4
pydantic-settings
python-multipart
Pillow
pytesseract>=0.3.10

# AI & Processing
sentence-transformers>=2.5.1
transformers>=4.38.2
torch>=2.2.1
spacy>=3.7.4

# OSINT Data Handling
beautifulsoup4>=4.12.3
requests>=2.31.0
feedparser>=6.0.11

# Task Queue
celery>=5.3.6
redis>=5.0.3

# Vector & Graph Support
neo4j>=5.18.0
pinecone-client>=3.1.0
